feat: migrate marking as read
