+++ package/src/components/Channel/hooks/useCreateChannelContext.ts
@@ -15,7 +15,6 @@ export const useCreateChannelContext = ({
   loadChannelAroundMessage,
   loadChannelAtFirstUnreadMessage,
   loading,
-  markRead,
   maxTimeBetweenGroupedMessages,
   maximumMessageLimit,
   reloadChannel,
@@ -40,7 +39,6 @@ export const useCreateChannelContext = ({
       loadChannelAroundMessage,
       loadChannelAtFirstUnreadMessage,
       loading,
-      markRead,
       maximumMessageLimit,
       maxTimeBetweenGroupedMessages,
       reloadChannel,
